Quiet new warnings.

+++ HexFiend/HFRepresenterTextView.m
@@ -117,7 +117,7 @@ static const NSTimeInterval HFCaretBlinkFrequency = 0.56;
         CGFloat advancePerColumn = [self advancePerColumn];
         HFASSERT(advancePerColumn > 0);
         CGFloat floatColumn = insetX / advancePerColumn;
-        HFASSERT(floatColumn >= 0 && floatColumn <= NSUIntegerMax);
+        HFASSERT(floatColumn >= 0 && floatColumn <= (CGFloat)NSUIntegerMax);
         CGFloat startOfColumn = advancePerColumn * HFFloor(floatColumn);
         HFASSERT(startOfColumn <= insetX);
         CGFloat xOffsetWithinColumn = insetX - startOfColumn;
@@ -127,7 +127,7 @@ static const NSTimeInterval HFCaretBlinkFrequency = 0.56;
         byteIndexWithinRow = bytesPerColumn * (NSUInteger)floatColumn + byteIndexWithinColumn; //this may trigger overflow to the next column, but that's OK
         byteIndexWithinRow = MIN(byteIndexWithinRow, bytesPerLine); //don't let clicking to the right of the line overflow to the next line
     }
-    HFASSERT(floatRow >= 0 && floatRow <= NSUIntegerMax);
+    HFASSERT(floatRow >= 0 && floatRow <= (CGFloat)NSUIntegerMax);
     NSUInteger row = (NSUInteger)floatRow;
     return (row * bytesPerLine + byteIndexWithinRow) / bytesPerCharacter;
 }
@@ -1066,7 +1066,7 @@ static size_t unionAndCleanLists(CGRect *rectList, __unsafe_unretained id *value
     }
     reverse <<= s; // shift when v's highest bits are zero
     
-    CGFloat hue = (CGFloat)reverse / ((CGFloat)1. + NSUIntegerMax);
+    CGFloat hue = (CGFloat)reverse / ((CGFloat)1. + (CGFloat)NSUIntegerMax);
 #if TARGET_OS_IPHONE
     return [UIColor colorWithHue:hue saturation:1. brightness:(CGFloat).6 alpha:alpha];
 #else
@@ -1810,7 +1810,7 @@ static size_t unionAndCleanLists(CGRect *rectList, __unsafe_unretained id *value
 - (NSUInteger)availableLineCount {
     CGFloat result = (CGFloat)ceil(CGRectGetHeight([self bounds]) / [self lineHeight]);
     HFASSERT(result >= 0.);
-    HFASSERT(result <= NSUIntegerMax);
+    HFASSERT(result <= (CGFloat)NSUIntegerMax);
     return (NSUInteger)result;
 }
